requirements.txt
# ==== CORE ====
numpy==1.26.4
scipy==1.11.4
timm>=0.9.12,<1.1
einops>=0.7.0
yacs>=0.1.8
PyYAML>=6.0.1
tqdm>=4.66.1
pillow>=10.2.0
packaging>=23.2
torchgeo

# ==== LOGGING / VISUALIZATION ====
tensorboard>=2.14
matplotlib>=3.8,<3.9
wandb>=0.17.0

# ==== VISION / EVAL ====
opencv-python-headless==4.10.0.84
scikit-learn>=1.3.0
kornia>=0.7.2
